Spec coverage for new service class
This involved migrating most of the request specs over to this file, with a little massaging.

Also includes RuboCop fixes.
