+++ app/models/user_preference.rb
@@ -13,8 +13,9 @@ class UserPreference < ActiveRecord::Base
   scope :for_account, ->(account_id) { where(account_id: account_id) }
 
   def self.for_preference_and_account(account_id, preference_codes)
-    joins(:preference)
-    .merge(Preference.with_code(preference_codes))
-    .for_account(account_id)
+    UserPreference
+      .joins(:preference)
+      .merge(Preference.with_code(preference_codes))
+      .for_account(account_id)
   end
 end
